Use a 2d array of OptionTableItem for the option hash table
Previously, a vector of OptionTableItem was used to represent
a bucket in the hash table, and the table consisted of an array
of such vectors representing buckets.

Unfortunately, older build compilers with incomplete C++11 support
may not have support for using initializer lists for initializing
std::vector, which is a non-aggregate type.

Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>