Introduce a temporary solution for querying options using current API
options-gen now generates the body of a swith statement that returns
the pointer to member in the new CompilerOptions class based on the
TR_CompilationOptions enum value. This opens up a way to remove the
need to replace all instances of boolean option querying using
getOption(TR_CompilationOptions o), simplifying the transition phase.

Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>