Initialize/shutdown portlib in JitBuilder and Test Compiler
This is one of the first steps towards enabling the use of the
port lib in the OMR Compiler. The responsibilty of initializing
and shutting down the port lib is assigned to the JitBuilder
and Test Compiler initialization phases, as they do not currently
run as part of a VM that uses the OMR Port Library.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>